requirements.txt
python-decouple==3.8
requests==2.31.0
mysql-connector-python==8.4.0